ml_server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
kafka-python-ng==2.2.3
redis==5.0.4
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.1
pytest==8.2.2
